# Optimized setbytes-functions