--- a/uint256.go
+++ b/uint256.go
@@ -111,7 +111,7 @@ func (z *Int) Uint64() uint64 {
 
 // Uint64WithOverflow returns the lower 64-bits of z and bool whether overflow occurred
 func (z *Int) Uint64WithOverflow() (uint64, bool) {
-	return z[0], z[1] != 0 || z[2] != 0 || z[3] != 0
+	return z[0], (z[1] | z[2] | z[3]) != 0
 }
 
 // Clone creates a new Int identical to z