@holiman 2020-06-04
Oh - Using a switch:
```
BenchmarkSetBytes/generic-6         	 4944483	       216 ns/op	       0 B/op	       0 allocs/op
BenchmarkSetBytes/specific-6        	10182189	       118 ns/op	       0 B/op	       0 allocs/op
```