@holiman 2020-06-04
Pushed a naive version now, along with a failing testcase as a reminder to fix it :)