@chfast 2020-06-04
> Not sure -- are you saying we should copy-paste the remaining functions from binary.BigEndian too? I'd rather not, if there's no gain in it.

Both variants generate the same assembly (https://godbolt.org/z/wJE3K_):
```go
func bigEndianUint40_1(b []byte) uint64 {
	_ = b[4] // bounds check hint to compiler; see golang.org/issue/14808
	return uint64(b[4]) | uint64(b[3])<<8 | uint64(b[2])<<16 | uint64(b[1])<<24 |
		uint64(b[0])<<32
}

func bigEndianUint40_2(b []byte) uint64 {
	_ = b[4] // bounds check hint to compiler; see golang.org/issue/14808
	return uint64(binary.BigEndian.Uint32(b[1:5])) | uint64(b[0])<<32
}
```
So there are 3 options here:
1. Keep it as it is (mixture of `binary.BigEndian` and single byte shifts).
2. Use `binary.BigEndian` also in `bigEndianUintXX` helpers.
3. Add more helpers based on single byte shifts - import of `"encoding/binary"` not needed.