@holiman 2020-06-04
> So there are 3 options here:
>

I don't really have a strong opinion... All else being equal, I think we should take whichever looks cleanest, so that the code is easy to understand and easy for a reviewer to vet.

I personally think `bigEndianUint40_1` is easier to understand than `bigEndianUint40_2`. I don't think we should explicitly _avoid_ `binary` (option 3), since it's a (hopefully) very stable native dependency.