@holiman 2020-06-04
I guess, if we want to align with `big.Int`, then we should stick with `master`. Then `SetBytes` is identical to `bigInt.SetBytes()` followed by conversion to `uint256.Int`.